requirements.txt
Pillow==10.1.0
Numpy==1.26
matplotlib==3.8.0